Successfully Writes to and Reads from Text Files
Much of this code was from the internet.  But this can be a baseline
for the other organization stuff for the app.  (I used the internet
because I tried to figure it on my own for like an hour and wasn’t
getting anywhere because of “IOException” not being recognized by the
compiler???)