Reading and printing text written onto txt files
JUST MIND THE .JAVA FILES (it wouldn’t let me commit otherwise?) The
only difference between this and the last update is that this one
can read text from the text file.  Right now I just have the text being
printed onto the screen, but this allows me to move onto storing text…
Credits for code from the internet can be found in the FileRead class.